requirements.txt
Django==2.1.7
django-extra-views==0.12.0
pytz==2018.9
six==1.12.0
